[BE] refactor: Service 통합 (#22)
